Redo behavior fragment (previously naming and workflow)
