+++ app/src/main/res/navigation/preferences.xml
@@ -2,32 +2,24 @@
 <navigation xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     android:id="@+id/preferences"
-    app:startDestination="@id/settings_fragment">
+    app:startDestination="@id/preferences_fragment">
 
     <include app:graph="@navigation/metadata"/>
 
     <fragment
-        android:id="@+id/settings_fragment"
-        android:name="org.phenoapps.intercross.fragments.SettingsFragment"
+        android:id="@+id/preferences_fragment"
+        android:name="org.phenoapps.intercross.fragments.preferences.PreferencesFragment"
         android:label="@string/settings_label">
 
-        <action android:id="@+id/action_from_settings_to_profile_fragment"
+        <action android:id="@+id/action_from_preferences_fragment_to_profile_fragment"
             app:destination="@+id/profile_preference_fragment"
             app:launchSingleTop="true"/>
 
-        <action android:id="@+id/action_from_settings_to_naming_fragment"
-            app:destination="@+id/naming_preference_fragment"
-            app:launchSingleTop="true"/>
-
-        <action android:id="@+id/action_from_settings_to_workflow_fragment"
-            app:destination="@+id/workflow_preference_fragment"
-            app:launchSingleTop="true"/>
-
-        <action android:id="@+id/action_from_settings_to_printing_fragment"
+        <action android:id="@+id/action_from_preferences_fragment_to_printing_fragment"
             app:destination="@+id/printing_preference_fragment"
             app:launchSingleTop="true"/>
 
-        <action android:id="@+id/action_from_settings_to_database_fragment"
+        <action android:id="@+id/action_from_preferences_fragment_to_database_fragment"
             app:destination="@+id/database_preference_fragment"
             app:launchSingleTop="true"/>
 
@@ -35,7 +27,7 @@
 <!--            app:destination="@+id/brapi_preference_fragment"-->
 <!--            app:launchSingleTop="true"/>-->
 
-        <action android:id="@+id/action_from_settings_to_about_fragment"
+        <action android:id="@+id/action_from_preferences_fragment_to_about_fragment"
             app:destination="@+id/about_preference_fragment"
             app:launchSingleTop="true"/>
 
@@ -44,21 +36,14 @@
             app:launchSingleTop="true" />
 
         <action
-            android:id="@+id/action_from_settings_to_naming_workflow_fragment"
-            app:destination="@id/naming_workflow_preference_fragment"
+            android:id="@+id/action_from_preferences_fragment_to_behavior_preferences_fragment"
+            app:destination="@id/behavior_preferences_fragment"
             app:launchSingleTop="true"/>
     </fragment>
 
     <fragment android:id="@+id/profile_preference_fragment"
         android:name="org.phenoapps.intercross.fragments.preferences.ProfileFragment"
         android:label="@string/preferences_profile_fragment_label"/>
-
-    <fragment android:id="@+id/workflow_preference_fragment"
-        android:name="org.phenoapps.intercross.fragments.preferences.WorkflowFragment"
-        android:label="@string/preferences_workflow_fragment_label">
-        <action android:id="@+id/action_from_workflow_to_metadata_list"
-            app:destination="@+id/metadata_navigation"/>
-    </fragment>
 
     <fragment android:id="@+id/printing_preference_fragment"
         android:name="org.phenoapps.intercross.fragments.preferences.PrintingFragment"
@@ -77,24 +62,15 @@
 <!--        android:name="org.phenoapps.intercross.fragments.preferences.BrapiFragment"-->
 <!--        android:label="@string/preferences_brapi_fragment_label"/>-->
 
-    <fragment android:id="@+id/naming_preference_fragment"
-        android:name="org.phenoapps.intercross.fragments.preferences.NamingFragment"
-        android:label="@string/preferences_naming_fragment_label">
-        <action
-            android:id="@+id/action_to_pattern_fragment"
-            app:destination="@id/pattern_fragment"
-            app:launchSingleTop="true" />
-    </fragment>
-
     <fragment
-        android:id="@+id/naming_workflow_preference_fragment"
-        android:name="org.phenoapps.intercross.fragments.preferences.NamingWorkflowFragment"
-        android:label="Naming and Workflow">
+        android:id="@+id/behavior_preferences_fragment"
+        android:name="org.phenoapps.intercross.fragments.preferences.BehaviorPreferencesFragment"
+        android:label="@string/prefs_behavior_title">
         <action
-            android:id="@+id/action_namingWorkflowFragment_to_patternFragment"
+            android:id="@+id/action_behavior_preferences_fragment_to_pattern_fragment"
             app:destination="@id/pattern_fragment" />
         <action
-            android:id="@+id/action_namingWorkflowFragment_to_metadataList"
+            android:id="@+id/action_behavior_preferences_fragment_to_metadataList"
             app:destination="@id/metadata_navigation" />
     </fragment>
