--- app/src/main/res/xml/naming_preferences.xml
@@ -1,24 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<PreferenceScreen
-    android:key="@string/root_naming"
-    xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto" >
-    <PreferenceCategory app:iconSpaceReserved="false">
-        <SwitchPreference
-            android:defaultValue="false"
-            android:icon="@drawable/ic_setting_male_blank"
-            android:key="@string/key_pref_naming_blank_male"
-            android:title="Allow blank male ID" />
-        <SwitchPreference
-            android:defaultValue="false"
-            android:icon="@drawable/ic_setting_cross_order"
-            android:key="@string/key_pref_naming_cross_order"
-            android:summary=""
-            android:title="Scan male first" />
-        <Preference
-            android:icon="@drawable/ic_setting_pattern_create"
-            android:key="@string/key_pref_naming_create_pattern"
-            android:title="Create cross ID pattern"
-            android:summary="UUID"/>
-    </PreferenceCategory>
-</PreferenceScreen>
